refactor: 파일 위치 변경 및 클래스 위치 변경
