--- app/src/main/java/woowacourse/movie/domain/data/MovieListData.kt
@@ -1,3 +1,0 @@
-package woowacourse.movie.domain.data
-
-interface MovieListData
