+++ app/src/main/java/woowacourse/movie/view/viewmodel/MovieListData.kt
@@ -1,0 +1,5 @@
+package woowacourse.movie.view.viewmodel
+
+sealed interface MovieListData {
+    data class ADData(val id: Int) : MovieListData
+}
